Simplified display size
Previously, the maximum display size was set by `DISPLAY_WIDTH_MAX` and
`DISPLAY_HEIGHT_MAX`, and a subset of this was used which was determined
by `width` and `height`, which were set in the Display() constructor.
Now, there is only `DISPLAY_WIDTH` and `DISPLAY_HEIGHT`, so the display
size cannot be changed at all. This makes things much simpler though.
Additionally, in Chronact.cpp, the window size is now determined by the
display size, which is read with display->GetPixelWidth() and
display->GetPixelHeight().