WriteText mask supports multiple tiles
The mask is now an array. The array is filled with ints which represent
the various tiles what will used, e.g. TILE_FADE_DARK. This is
essentially a colour pallet, but for tiles. Each number in the text then
relates to a tile in the mask array. 1 = first tile, 2 = second tile,
etc. This allows a mask array to have up to 9 different tiles used in a
pallet. Anything outside of the numbers 1 to 9 are ignored.