Object structure and rendering
There is one abstract base class for objects, called GameObject::Base.
From this are two children, Dynamic and Static. All game objects must
then inherit from Dynamic or Static. The door object inherits from
Static, because it does not move. The player object inherits from
Dynamic, because it does move. The class LevelData stores a vector of
GameObject::Base* and controls the creation and deletion of all game
objects. LevelManager creates an instance of LevelData and calls the
corresponding functions to retrieve information about object positions
and display characters, and renders this on top of the basic level
layout (made up of the walls and floors). In this way objects are always
rendered on top of the level.