Player can move with keyboard input
The GameObject::Dynamic class was extended to add some basic movement
functions. This included checking whether the movement was within the
bounds of the level. For this to occur, the object needed to know the
dimensions of the level. This meant it required a pointer to the
GameEngine (which in turn has a pointer to Display, which keeps track of
the display size (though in hindsight it should actually read the size
from LevelData or LevelManager)). Anyway, now a pointer to GameEngine
has been passed all the way down to each GameObject, allowing it to
access GameEngine stuff (which could include things like the global
time). The Player class was then extended to act upon some simple
keyboard input and move itself.