Fullscreen and scaling support, and better config
Added support for playing in fullscreen through setting "fullscreen=yes"
in config.txt. Added support for scaling (but only in fullscreen). By
default the scaling is at 1:1, so in fullscreen mode there will be a
thick black border all around. The scaling is adjusted by setting
"fullscreen_scale_x" and "fullscreen_scale_y" in config.txt. Improved
the support for config.txt so that it reads values properly, and can
read floats.