Improved object inheritance model
A base constructor has been placed in Dynamic and Static. Each object
that is created that inherits from these classes then calls their
constructors first. Instead of passing a reference to levelData, a
reference to levelManager is passed. Since levelManager has a reference
to levelData, the same data can still be accessed. In additiona,
levelManager now has an instance of iterationData, which can also be
accessed by all objects, and which stores information about the current
time travel situation (e.g. which iteration we are in). The timline is
also updated with a T-x counter.