Actions and tag based object collision detection
The player can press Z to interact. For this to be useful, collision
detection was added with against objects. All objects that want to be
identified must assign themselves a tag (in LevelData.cpp). For ease of
reference, tags are defined in Tags.hpp, though this is not strictly
necessary.