Can get in time machine and go back in time
Added variables canHaveControl and hasControl to GameObject::Base. These
are used to ensure that only one object at a time is being controlled by
the player, and to ensure that control is not passed to an object which
cannot assume control. The function Controlling() returns whether the
object in question is currently in control. The player can now go to the
time machine and pass control to it, and use it to go back in time.