A plus is displayed between each multiple object
Since when multiple objects are on the same tile, if they are all
identical, switching between them will not actually indicate that there
are multiple objects. So now, a plus sign is displayed between each
multiple object on the same square in order to show that it is cycling
between all the objects.