Continuous keyboard input without bugs
Previously, each object read controls directly from sf::Keyboard. Now,
all objects must read controls through the Controls class. This then
allows an object to ensure that if it acts upon a key press in one
frame, that another object cannot act upon the key press in the same
frame. For example, the player presses Z to enter a time machine, but Z
also goes back in time while in a time machine. With the old system, a
player who entered a time machine would automatically go back in time 1
unit because the key press event that was registered for the action of
getting into the time machine also triggered the event for going back in
time. This could lead to time related bugs because the order of
execution is arbitrary (and depends upon the order in which the objects
were created). Now, if the player presses Z to get in to a time machine,
the code then resets a key delay function which then disallows the time
machine from acting upon any key press events until a certain delay has
passed (currently defined by MOVEMENT_DELAY in Defs.hpp). This delay is
optional, and is ensured by calling the function
Controls::GetKeyDelaySufficient(), which returns true if it has been
long enough to act up on another key press. The delay is reset with
Controls::ResetKeyDelay(). It should be reset whenever a delay is wanted
to be enforced between now and the next key input.