Detection of whether timeData exists
Added function Dynamic::TimeDataExists() which returns true if timeData
exists for the object for the current time. This should be called before
Dynamic::TimeDataRead() else it may return invalid (default) timeData.
To show this works, the Player is moved to -1,-1 when no timeData exists
which means it is not rendered.