Fixed expiryTime bug
The expiryTime for the first player object (the original) was left
undefined in the constructor. It was meant to have a value of -1 set.
This meant that during the game, it was possible for the original player
object to be triggered by its expiryTime, if it happened to land between
a value of 0 and 40, since that is the time limit of the game at the
moment. This would sometimes cause a crash because "true" would be
written into an undefined pointer location as a result, and this left
the game in a situation where there were no active players.