Allowed object coordinates outside of level
Objects can now have coordinates set outside of the level (e.g. -1,-1)
which will effectively mean they are not rendered. This is allowed now
that LevelManager::UpdateDisplay checks and ignores all coordinates
outside of the level. If this step did not occur, the game will most
likely crash due to memory corruption or the likes.