TimeData includes whether the object exists
In some cases, an object may write TimeData when the object itself does
not exist in that time frame (e.g. when a clone passes control back to
its parent, and stops existing, but writes TimeData). As such, TimeData
now writes whether the object exists in that time, and as such, the
presence of TimeData should no longer be used to detect whether the
object exists.