Applied Event::DoorOpen
Adjusted the ordering of when events occur in
EventData::ExecuteBackwardsEvents so that backwards events are occured
in backwards order. I think this is necessary. Also, rather than
creating an event and calling it immediately, an object creates an event
and it is called automatically when the time changes forwards. This
works because forward events have been set to occur 1 time unit after
they are registered to occur (whereas before, backward events were set
to occur 1 time unit before they were registered to occur). Things are
getting slightly confusing.