Fixed player expiry/transfer paradox
A paradox occurred because a clone would create an Event to move to a
new location just before it expires, so the event would be created but
then it would never reach the location because expired objects are moved
to location -1,-1. This caused the Event to cause a paradox any time it
was revisited. Maybe an expiry event needs to be added?