Fixed order of events paradox
A paradox used to occur when a player expired due to the order in which
the events were processed. Now, the order has been corrected so that the
backwards events occur before the forwards events, and time does not
actually change until after the forward events/before the backward
events. So `time` in IterationData has been made a public variable
because it is now modified by LevelManager at the appropriate point.