Links are read from the level file
LevelManager has been extended a great deal to read the level file and
generate Links, with error checking. This adds a massive deal of
flexibility to level creation because any object that inherits from
GameObject::StaticLinkable (which so far includes two types of Door and
a Lever) can be linked together in any way.