Fixed player movement paradox
A paradox was detected in the following situation: a player goes back in
time, moves around a bit, and then goes back in time again the instant
they expire. The reason for this was because in
Player::UpdateTimeChanged() there was some old code which executed when
the player expired which moved it to position -1,-1. This was left over
from the old time travel system, before events were added. This code
caused a paradox because it meant the player was not in the expected
position when the event ran which moved the player, and relied upon the
player being in a certain position.