Timeline displays relation of clones
On the right side of the timeline is now displayed the designation of
the parent of the clone on that row. For the original player, who has no
parent, no parent designation is shown. The timeline has also been
slightly modified to be less cluttered with symbols at either end. In
addition, level 002 was adjusted slightly so that the critical time is
equal to the time limit.