Player can be crushed to death by doors
Added Event::PlayerDie_Linkable and applied it in the form of doors
killing the player. When the player is crushed by a door, it takes up 1
time unit, and automatically steps forward in time. A dead player is
represented by a heart. To show the controlling player is dead on the
timeline, it switches between displaying the heart for 0.75 seconds and
the controlling player for 0.25 seconds.