Added LevelData::GetAllPlayersOnShutDoors()
This function was added in an attempt to fix the bug whereby the
controlling player cannot kill another player directly with doors. This
is because the death event is only added if the player dying is the
player in control (to avoid duplicate events). The new approach that was
taken was to, when a lever is pulled, look forward one frame and see if
any players should be killed as a result of the lever. This solution is
flawed logically because we do not know where the players will be next
frame! It works only if the player being killed remains stationary
(because then their position remains constant and so we know they will
die). Since we do not know if a player will stay stationary, we cannot
use this solution (see commented out code in Player).