Added tentative events
Tentative events are used to create an event for an event that may or
may not occur next step. They have been applied in the form of killing a
player who gets crushed by a door when a lever is pulled. A tentative
even is created for the door controlled by the lever for killing every
single player. In the next step, all the tentative events are tested,
and all those which do not succeed are deleted. Those which do succeed
are added to the list of proper events. One major problem with this is
that we cannot create a tentative event to kill someone who does not
exist yet, so players cannot be killed by previous iterations of
themselves (yet).