Fixed Issue #3: Players are unaffected by a door shutting on them
Pressure plates can now cause death via doors. Levers can also cause
death via doors (which was added over the last few commits). Added the
bool `failOnPlayerPos` in Event::Base which controls whether
ForwardEvent() and BackwardEvent() will fail if the player is in the
wrong position. This is used for when an event is tested in
Player::CreateFutureRepercussions(). The level 003 was modified to be a
suitable testing level for killing players with doors operated by a
lever and a pressure plate.