Fixed level changing crash
Changing a level by changing the state from within a state seemed to
cause a crash. This has been fixed by introducing the functions
GameEngine::LoadLevelNext() and GameEngine::LoadLevelCurrent(), which
set flags that are checked at the end of each cycle. This means the
level can switch at a more appropriate time, and cause no crash. The
side effect of this is that one must realise calling
game->LoadLevelNext() will not change the level instantly, but will run
until the end of the current frame first.