Added title to the outline structure
This means that the title can be loaded at arbitrary points in the
outline, or not at all. More importantly, it means testing is much
easier because the first item in the content outline can be set to a
level, which will be loaded immediately - rather than before where you
had to sit through the title screen each time. This required adding a
new state, EntryState, whose sole purpose is to load the next
appropriate state.