Added event ordering to fix door death bug
Found a bug where if a players move on to a door the frame it closes,
they do not die. This was due to the order of events - the door shut and
then the player moved. Event ordering was added, so now the door close
death event always occurs after player movement, meaning the death
occurs.