Added GameObject::MotionAlarm
The motion alarm is essentially a pressure plate that cannot be
deactivated. The moment a player steps into the motion alarm square, it
is activated, and stays active. It also has an animation while active,
similar to how the time machine has an animation. In fact, it borrows
the same #define for controlling its animation speed.