Works in OS X Mavericks with Xcode
Had to include the cmath library for use of sin and cos functions. Had
to change code of StringToInt slightly, which is probably for the best.
Had trouble with a variable length int vector in LevelManager.cpp, so
for a temporary fix I have set a constant upper limit of 9999.