Removed delay skip on key release
Removed the feature added in the previous commit where if a key was
released, then a key pressed soon would not be subject to the longer
delay. It was removed because the controls felt rather awkward and
unpredictable with their timings. The root of this problem seems to be
that there is just a global key delay timer, rather than one for each
key.