Applied constant upper limit to UpdateDisplay
Since compiling under clang (rather than gcc) there was an issue with
the vector size being defined at run time. In order to fix this, the
upper limit needed to be calculated (instead of the exact size). This
was done by using the constants that define the maximum dimensions of
the room.